Feature#011/armies (#28)
# Army Feature

The army feature introduces the ability to raise armies from claimbuilds, move armies around the map, bind and unbind characters from them, and much more.

## New commands and their functionality

### General commands
- `/create army`
Raises a new army from an existing claimbuild. Every military claimbuild can have one active army, with the exception of capitals, which can have two.
- `/disband army-or-company`
Disbands an army or company. Can only be done by faction leaders or lords.
- `/bind army-or-company`
Binds an army to an existing Player/Roleplay Character. The army has to be of the same faction as the player or one of its allies. Note that the army must be located in the same region as the player. Faction leaders/Lords can also bind other players of the same faction.
- `/unbind army-or-company`
Unbinds a player from an army. Leaders/Lords can unbind other players of the same faction.
- `/move army-or-company`
Creates a movement for an army or company. The shortest path to the destination is calculated by the application. Armies/Companies that were created less than 24 hours ago are unable to move.
- `/cancel-move army`
Cancels an ongoing army/company movement. The army/company stays at the current location.
- `/station army-or-company`
Stations an army/company at a claimbuild. Armies/companies can only be stationed at allied claimbuilds. Stationed armies are unable to move.
- `/unstation army-or-company`
Unstations an army/company from a claimbuild. The army/company is able to move again.
- `/pick-siege`
Picks up siege equipment from a claimbuild. Only sieges available at the claimbuild can be picked up.
- `/heal-start army`
Starts healing an army that has lost units in a battle. The army needs to be stationed at a claimbuild that has a house of healing.
- `/heal-stop army`
Stops the healing of an army that is currently healing.
- `/get upkeep`
Prints out the amount of monthly upkeep a faction has to pay for their armies.
- `/get unpaid-armies-or-companies`
Prints out all the armies or companies that have no registered payment yet.

### Staff commands
The following commands can only be used by staff members and are mainly for maintainence.
- `/delete army-or-company`
Force-deletes an army or company without checks for faction leader permissions, etc.
- `/update paid`
Updates if an army is paid for or not.