Feature#026/missing commands (#35)
# Missing commands
In order for the 1.0 release to happen, we must implement some commands that have been missed by previous features.

## New commands and their functionality

### General commands
- `/injure-character`
Injures your roleplay character. Injured characters cannot be bound to armies. If your character is bound to an army and you use this command, it will get unbound from said army.
- `/heal-start character`
Starts healing an injured character. In order to heal your character needs to be in the same region as a claimbuild with a House of Healing. Healing characters cannot move or be bound to armies.
- `/heal-stop character`
Stops healing of your character.

### Staff commands
The following commands can only be used by staff members and are mainly for maintainence.
- `/update faction-leader`
Sets the faction leader of a faction. The specified player must be part of the inputted faction.