Added BattleService for creating new Battles
First iteration of the BattleService. Includes a method to create a
Battle - saves and returns a battle object containing the armies, time,
etc. after checking for conditions like:
Attacker is allowed to attack with army, if they are at war, if its a
fieldbattle / cb battle etc.

Missing: Adding the proper War object and times into the Battle object.
Need to add a util method to check if the army is in a 24h range and
therefore allowed to attack.

	Add:
	- BattleRepository.java
	- CreateBattleDto.java
	- BattleService.java
	- BattleServiceException.java

	Modify:
	- Battle.java

Issues: #19 - battles and wars
Signed-off-by: Sascha Hren <syncheleven@gmail.com>